+++ HISTORY.rst
@@ -21,6 +21,8 @@ History
 * Now employing PEP8 + Black compatible autoformatting.
 * Added Windows and macOS images to Travis CI build ensemble.
 * Added variable thresholds for tasmax and tasmin in daily_freezethaw_events.
+* Updated subset.py to use date formatted strings ("%Y", "%Y%m" etc.) in temporal subsetting
+
 
 0.10-beta (2019-06-06)
 ----------------------